refactor isHoliday by using it as a helper
