+++ frontend/src/views/ClockPage.vue
@@ -16,24 +16,18 @@
 </template>
 
 <script>
-import { ref, computed } from 'vue';
+import { ref } from 'vue';
 import { useStore } from 'vuex';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 import timezone from 'dayjs/plugin/timezone';
 import attendanceAPI from './../apis/attendances';
-import { Toast, storeCheck } from './../utils/helpers';
-
-import calendar from '../data/calendar.json';
+import { Toast, storeCheck, isHoliday } from './../utils/helpers';
 
 dayjs.extend(utc, timezone);
 
 export default {
   setup() {
-    const filteredCalendar = computed(() => {
-      return calendar.filter((entry) => entry.是否放假 === '2');
-    });
-
     const store = useStore();
 
     const clockedInValue = localStorage.getItem('clockedIn');
@@ -55,17 +49,6 @@ export default {
 
     // check if the user is clocked in
     const clockedIn = ref(clockedInCheck);
-
-    const isHoliday = filteredCalendar.value.some((entry) => {
-      const year = entry.西元日期.substring(0, 4);
-      const month = entry.西元日期.substring(4, 6);
-      const day = entry.西元日期.substring(6, 8);
-
-      const entryDate = dayjs(`${year}-${month}-${day}`).format(
-        'YYYY-MM-DD 00:00:00'
-      );
-      return date.value === entryDate;
-    });
 
     date.value = dayjs.utc().local().format('YYYY-MM-DD 00:00:00');
     store.commit('setDate', date.value);
@@ -202,7 +185,6 @@ export default {
       clockedIn,
       clockIn,
       clockOut,
-      filteredCalendar,
       isHoliday,
     };
   },
