+++ frontend/src/views/GPSPage.vue
@@ -1,15 +1,19 @@
 <template>
   <div class="container py-5">
-    <GPSMap />
+    <h2 v-if="isHoliday">今天放假，好好休息！</h2>
+    <div v-else><GPSMap /></div>
   </div>
 </template>
 
 <script>
 import GPSMap from './../components/GPSMap.vue';
+import { isHoliday } from './../utils/helpers.js';
 export default {
   components: {
     GPSMap,
   },
-  setup() {},
+  setup() {
+    return { isHoliday };
+  },
 };
 </script>
