+++ frontend/src/views/QRPage.vue
@@ -1,20 +1,24 @@
 <template>
   <main>
     <div class="container py-5">
-      <div v-if="currentUser.role === 'admin'">
-        <a href="#/clocking/qrcode/generate">產生QR Code</a>
-        | <a href="#/clocking/qrcode/read">讀取QR Code</a>
+      <h2 v-if="isHoliday">今天放假，好好休息！</h2>
+      <div v-else>
+        <div v-if="currentUser.role === 'admin'">
+          <a href="#/clocking/qrcode/generate">產生QR Code</a>
+          | <a href="#/clocking/qrcode/read">讀取QR Code</a>
+          <component :is="currentView" />
+        </div>
       </div>
-      <component :is="currentView" />
     </div>
   </main>
 </template>
 
 <script>
-import QRGenerator from '../components/QRGenerator.vue';
-import QRReader from '../components/QRReader.vue';
 import { computed, ref } from 'vue';
 import { mapState, useStore } from 'vuex';
+import QRGenerator from '../components/QRGenerator.vue';
+import QRReader from '../components/QRReader.vue';
+import { isHoliday } from '../utils/helpers';
 
 export default {
   setup() {
@@ -26,8 +30,6 @@ export default {
       resultStoreState[item] = computed(resFuc.bind({ $store: store }));
     });
     const { currentUser } = { ...resultStoreState };
-
-    console.log(currentUser);
 
     const routes = {
       '/clocking/qrcode/generate': QRGenerator,
@@ -41,11 +43,10 @@ export default {
     });
 
     const currentView = computed(() => {
-      console.log(currentPath.value.slice(1));
       return routes[currentPath.value.slice(1) || '/'] || 'NotFound';
     });
 
-    return { currentView, currentUser };
+    return { currentView, currentUser, isHoliday };
   },
 };
 </script>
